'added' the projectile trial, however it is causing Unity to crash for some reason.
If the Projectile trail is active unity crashes when I hit 'Play' If the Projectile is Disabled the game will play as normal. However if I turn on the Projectile trail mid game then Unity crashes. I am looking for a solution, but I don't Know whats going on. The very first frame that the Projectile trail is active it crashes.